requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyGithub==1.59.1
requests==2.31.0
pylint==3.0.3
bandit==1.7.5
cryptography